strongafter-assistant/backend/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
python-dotenv==1.0.1
gunicorn==21.2.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0
google-generativeai==0.3.2 
google-cloud-aiplatform==1.42.1
vertexai==0.0.1
openai==0.28.1
numpy==1.26.4
faiss-cpu==1.7.4 